requirements.txt
Django==5.1.4
djangorestframework
psycopg
redis
django-cors-headers
python-dotenv
gunicorn
django-filter
pytest
pytest-django
black
isort
flake8
drf-spectacular
drf-spectacular-sidecar
djangorestframework-simplejwt